+++ Package.swift
@@ -3,12 +3,9 @@
 import PackageDescription
 
 let package = Package(
-    name: "BroadAppsIOSPlatform",
+    name: "BroadPlatformIntegration",
     platforms: [
         .iOS(.v17),
-    ],
-    products: [
-        .library(name: "BroadUIFlows", targets: ["BroadUIFlows"]),
     ],
     dependencies: [
         .package(
@@ -24,19 +21,14 @@ let package = Package(
             exact: "1.0.0"
         ),
         .package(
+            url: "https://github.com/BroadApps-official/broad-ui-flows-ios.git",
+            exact: "1.0.0"
+        ),
+        .package(
             url: "https://github.com/Swinject/Swinject.git",
             exact: "2.10.0"
         ),
     ],
-    targets: [
-        .target(
-            name: "BroadUIFlows",
-            dependencies: [
-                .product(name: "BroadCore", package: "broad-core-ios"),
-                .product(name: "BroadMonetization", package: "broad-monetization-ios"),
-                .product(name: "Swinject", package: "Swinject"),
-            ]
-        ),
-    ],
+    targets: [],
     swiftLanguageModes: [.v5]
 )