+++ Sources/BroadUIFlows/Domain/AppFlow/AppFlowRoute.swift
@@ -1,6 +1,0 @@
-public enum AppFlowRoute: Equatable, Sendable {
-    case launch
-    case onboarding
-    case initialPaywall
-    case main
-}